pyproject.toml
[tool.poetry]
name = "Financial Control API"
version = "0.1.0"
description = ""
authors = ["Mateus Gomes <mateusgomes.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = { extras = ["standard"], version = "^0.115.5" }
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.19"
pyjwt = "^2.10.1"
tzdata = "^2024.2"
psycopg = {extras = ["binary"], version = "^3.2.3"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
isort = "^5.13.2"
pylint = "^3.3.1"
black = "^24.10.0"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"
testcontainers = "^4.8.2"


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.isort]
profile = "black"
line_length = 79
skip = ["migrations"]

[tool.black]
line-length = 79
exclude = "/migrations/"

[tool.pylint]
ignore = ["migrations"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
